requirements.txt
asgiref==3.2.7
boto3==1.13.25
botocore==1.16.25
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-heroku==0.3.1
django-sendgrid-v5==0.8.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.6.0
jmespath==0.10.0
Markdown==3.2.2
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
s3transfer==0.3.3
sendgrid==6.3.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
zipp==3.1.0
